COM-1754 - Add Mixin
